rename trainer and update in dependencies,remove old
